fix: adds rollup bundle to package for cdn and reorganizes some project files
Adds rollup bundle to dist for cdn usage. Reorganizes project files. Adds excluded files to
tsconfig.prod.json.
