--- .travis.yml
@@ -1,38 +1,0 @@
-language: node_js
-os: linux
-dist: xenial
-
-node_js:
-  - 14
-  - 12
-  - 10
-
-script:
-  - yarn lint
-  - yarn test
-  - yarn build
-
-cache: yarn
-jobs:
-  include:
-    - stage: release
-      node_js: lts/*
-      before_deploy: yarn build
-      deploy:
-        provider: script
-        skip_cleanup: true
-        script:
-          - npx semantic-release
-    - stage: docs
-      cache:
-        directories: node_modules
-      script: yarn docs:build
-      deploy:
-        provider: pages
-        skip_cleanup: true
-        local_dir: docs/.vuepress/dist
-        github_token: $GH_TOKEN # A token generated on GitHub allowing Travis to push code on you repository. Set in the Travis settings page of your repository, as a secure variable
-        repo: AJStacy/Adze
-        keep_history: true
-        on:
-          branch: master
