+++ demo.ts
@@ -1,6 +1,6 @@
 import { JSDOM } from 'jsdom';
 
-import * as AdzeLib from './dist';
+import * as AdzeLib from './src';
 import runDemo from './demo_funcs.js';
 
 const dom = new JSDOM(`<!DOCTYPE html><div id="test">Hello world</div>`);
