+++ package.json
@@ -13,7 +13,9 @@
   "scripts": {
     "prebuild": "npm run clean",
     "build": "tsc --project tsconfig.prod.json",
+    "postbuild": "npm run bundle",
     "clean": "rm -rf ./dist",
+    "bundle": "rollup ./dist/index.js --file ./dist/bundle.js --format iife",
     "commit": "cz",
     "prepare": "husky install",
     "prepack": "npm run build",
@@ -59,6 +61,7 @@
     "jsdom": "^16.3.0",
     "lint-staged": "^10.5.4",
     "prettier": "2.2.1",
+    "rollup": "^2.45.2",
     "semantic-release": "^17.4.2",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3",
