+++ src/util/index.ts
@@ -1,3 +1,5 @@
-export * from './util';
-export * from './type-guards';
+export * from './bundle';
 export * from './data';
+export * from './shed';
+export * from './type-guards';
+export * from './util';
