+++ tsconfig.prod.json
@@ -1,4 +1,5 @@
 {
   "extends": "./tsconfig",
-  "include": ["src/**/*.ts"]
+  "include": ["src/**/*.ts"],
+  "exclude": ["node_modules", "./ava.config.js", "./demo_funcs.js", "./demo.ts"]
 }
